binder/Dockerfile: fix execution on mybinder.org
The recent image update was failing to build on MyBinder, seemingly
due to a bug that happens when older versions of Docker are combined
with newer Ubuntu releases:

https://stackoverflow.com/questions/66319610/gpg-error-in-ubuntu-21-04-after-second-apt-get-update-during-docker-build
https://bugs.launchpad.net/ubuntu/+source/glibc/+bug/1916485

Dropping to Ubuntu 20.10 doesn't work because it's EOLed (I think -- the
`apt update` errors out), so we have to drop to 20.04. That in turn
means that we need to install Node.js manually since 20.04's version of
it is too old for JupyterLab.